+++ packages/react-router-dom-v5-compat/index.ts
@@ -85,6 +85,7 @@ export type {
   Params,
   Path,
   PathMatch,
+  PathParam,
   PathPattern,
   PathRouteProps,
   Pathname,