+++ packages/react-router-dom/index.tsx
@@ -126,6 +126,7 @@ export type {
   Path,
   PathMatch,
   Pathname,
+  PathParam,
   PathPattern,
   PathRouteProps,
   RedirectFunction,